cleanup cast

+++ tt_metal/llrt/wormhole/wh_hal.cpp
@@ -73,7 +73,7 @@ void Hal::initialize_wh() {
     this->iram_relocate_func_ = [](uint64_t addr, HalProgrammableCoreType type) {
         if (std::getenv("TT_METAL_ENABLE_ERISC_IRAM")) {
             if (type == HalProgrammableCoreType::ACTIVE_ETH &&
-                addr == static_cast<uint64_t>(static_cast<uint32_t>(eth_iram_mem::address_map::ERISC_IRAM_BASE))) {
+                addr == static_cast<uint32_t>(eth_iram_mem::address_map::ERISC_IRAM_BASE)) {
                 // IRAM enabled program starts from ERISC_IRAM_BASE. This relocation is for where to put the program.
                 // At first the program is placed on ERISC_IRAM_BASE, then erisc.cc copies to local IRAM.
                 return (uint64_t)eth_l1_mem::address_map::KERNEL_BASE;
